reimplemented ConnectionMConstraint

+++ common/src/main/kotlin/net/spaceeye/vmod/toolgun/modes/state/RopeMode.kt
@@ -25,7 +25,7 @@ import net.spaceeye.vmod.utils.RaycastFunctions
 
 class RopeMode: ExtendableToolgunMode(), RopeGUI, RopeHUD {
     var compliance: Double by get(0, 1e-20, { ServerLimits.instance.compliance.get(it) })
-    var maxForce: Double by get(1, 1e10, { ServerLimits.instance.maxForce.get(it) })
+    var maxForce: Float by get(1, 1e10f, { ServerLimits.instance.maxForce.get(it) })
     var fixedDistance: Double by get(2, -1.0, {ServerLimits.instance.fixedDistance.get(it)})
 
     var primaryFirstRaycast: Boolean by get(3, false)
